third_party/blink/web_tests/http/tests/security/local-CSS-from-remote-expected.txt

CONSOLE ERROR: Not allowed to load local resource: cssStyle.css
This test is to see if a remote file can include a local CSS style.
Currently this test cannot be run manually on Windows because we do not have a function like pathToLocalResource() outside of DRT.

If the background is yellow then the CSS was loaded.


Test Passed: Local CSS not remotely loaded.
